render linestringDem compatible with GdalExtractProfile
