--- zoo-project/zoo-services/linestring-dem/linestring-dem/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "gdal-async": "^3.5.1"
-  }
-}
