+++ zoo-project/zoo-services/linestringDem/.gitignore
