+++ zoo-project/zoo-services/linestringDem/0347.min.geojson
